+++ src/test/java/lotto/step1/domain/StringAddCalculatorTest.java
@@ -1,5 +1,6 @@
-package lotto.step1;
+package lotto.step1.domain;
 
+import lotto.step1.domain.StringAddCalculator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
